Add files via upload
Welcome to the Adidas Sales Analysis! In this project, we dive into the world of Adidas, as we know the adidas is well-known brand. We're exploring how much Adidas sold and earned in year 2020 and 2021 and also what people loved, where they shopped the most, and overall company’s performance. We're going to uncover sales trends and understand what made Adidas successful.